+++ frontend/src/pages/organization/ProductSettingsPage/SecretsManagement/project-templates/ProjectTemplatePage.tsx
@@ -1,11 +1,9 @@
-import { Helmet } from "react-helmet";
 import { useNavigate } from "@tanstack/react-router";
 
-import { PageHeader } from "@app/components/v2";
 import { useOrganization } from "@app/context";
 import { ProjectType } from "@app/hooks/api/projects/types";
 
-import { EditProjectTemplateSection } from "./EditProjectTemplateSection";
+import { ProjectTemplatePage as ProjectTemplatePageBase } from "../../project-templates/ProjectTemplatePage";
 
 type Props = {
   templateId: string;
@@ -27,23 +25,10 @@ export const ProjectTemplatePage = ({ templateId }: Props) => {
   };
 
   return (
-    <>
-      <Helmet>
-        <title>Project Template | Infisical</title>
-        <link rel="icon" href="/infisical.ico" />
-      </Helmet>
-      <div className="h-full">
-        <div className="mx-auto h-full w-full max-w-8xl bg-bunker-800 text-white">
-          <PageHeader
-            scope={ProjectType.SecretManager}
-            title="Project Template"
-            description="Configure project template roles, environments, and memberships."
-          />
-          <div className="pb-8">
-            <EditProjectTemplateSection templateId={templateId} onBack={navigateBack} />
-          </div>
-        </div>
-      </div>
-    </>
+    <ProjectTemplatePageBase
+      templateId={templateId}
+      projectType={ProjectType.SecretManager}
+      onBack={navigateBack}
+    />
   );
 };
