+++ frontend/src/pages/organization/ProductSettingsPage/project-templates/DeleteProjectTemplateModal.tsx
