+++ frontend/src/pages/organization/ProductSettingsPage/project-templates/EditProjectTemplateSection/EditProjectTemplateSection.tsx
@@ -1,7 +1,6 @@
 import { useMemo } from "react";
-import { ChevronLeft } from "lucide-react";
 
-import { Button, Empty, EmptyHeader, EmptyTitle, PageLoader } from "@app/components/v3";
+import { Empty, EmptyHeader, EmptyTitle, PageLoader } from "@app/components/v3";
 import { ProjectType } from "@app/hooks/api/projects/types";
 import { InfisicalProjectTemplate, useListProjectTemplates } from "@app/hooks/api/projectTemplates";
 
@@ -9,16 +8,14 @@ import { EditProjectTemplate } from "./components";
 
 type Props = {
   templateId: string;
-  onBack: () => void;
+  projectType: ProjectType;
 };
 
-export const EditProjectTemplateSection = ({ templateId, onBack }: Props) => {
+export const EditProjectTemplateSection = ({ templateId, projectType }: Props) => {
   const { data: projectTemplates, isPending } = useListProjectTemplates();
 
   const projectTemplate = useMemo(() => {
-    const template = projectTemplates?.find(
-      (t) => t.id === templateId && t.type === ProjectType.SecretManager
-    );
+    const template = projectTemplates?.find((t) => t.id === templateId && t.type === projectType);
     if (!template) return undefined;
 
     return {
@@ -28,7 +25,7 @@ export const EditProjectTemplateSection = ({ templateId, onBack }: Props) => {
         permissions: role.permissions.filter((perm) => perm.subject !== "secret-events")
       }))
     };
-  }, [projectTemplates, templateId]);
+  }, [projectTemplates, templateId, projectType]);
 
   const isInfisicalTemplate = Object.values(InfisicalProjectTemplate).includes(
     projectTemplate?.name as InfisicalProjectTemplate
@@ -36,15 +33,6 @@ export const EditProjectTemplateSection = ({ templateId, onBack }: Props) => {
 
   return (
     <div>
-      <Button
-        variant="ghost"
-        type="button"
-        onClick={onBack}
-        className="mb-6 px-0 text-mineshaft-300 hover:bg-transparent hover:text-mineshaft-100"
-      >
-        <ChevronLeft className="size-4" />
-        Back to Templates
-      </Button>
       {/* eslint-disable-next-line no-nested-ternary */}
       {isPending ? (
         <div className="flex h-[60vh] w-full items-center justify-center p-24">
@@ -54,7 +42,6 @@ export const EditProjectTemplateSection = ({ templateId, onBack }: Props) => {
         <EditProjectTemplate
           isInfisicalTemplate={isInfisicalTemplate}
           projectTemplate={projectTemplate}
-          onBack={onBack}
         />
       ) : (
         <Empty className="border border-dashed">
