+++ frontend/src/pages/organization/ProductSettingsPage/project-templates/EditProjectTemplateSection/components/EditProjectTemplate.tsx
@@ -1,0 +1,39 @@
+import { ProjectType } from "@app/hooks/api/projects/types";
+import { TProjectTemplate } from "@app/hooks/api/projectTemplates";
+
+import { ProjectTemplateEnvironmentsForm } from "./ProjectTemplateEnvironmentsForm";
+import { ProjectTemplateGroupsSection } from "./ProjectTemplateGroupsSection";
+import { ProjectTemplateIdentitiesSection } from "./ProjectTemplateIdentitiesSection";
+import { ProjectTemplateRolesSection } from "./ProjectTemplateRolesSection";
+import { ProjectTemplateUsersSection } from "./ProjectTemplateUsersSection";
+
+type Props = {
+  projectTemplate: TProjectTemplate;
+  isInfisicalTemplate: boolean;
+};
+
+export const EditProjectTemplate = ({ isInfisicalTemplate, projectTemplate }: Props) => {
+  const { type } = projectTemplate;
+
+  return (
+    <>
+      {type === ProjectType.SecretManager && (
+        <ProjectTemplateEnvironmentsForm
+          isInfisicalTemplate={isInfisicalTemplate}
+          projectTemplate={projectTemplate}
+        />
+      )}
+      <ProjectTemplateRolesSection
+        isInfisicalTemplate={isInfisicalTemplate}
+        projectTemplate={projectTemplate}
+      />
+      {!isInfisicalTemplate && (
+        <>
+          <ProjectTemplateUsersSection projectTemplate={projectTemplate} />
+          <ProjectTemplateGroupsSection projectTemplate={projectTemplate} />
+          <ProjectTemplateIdentitiesSection projectTemplate={projectTemplate} />
+        </>
+      )}
+    </>
+  );
+};
