+++ frontend/src/pages/organization/ProductSettingsPage/project-templates/EditProjectTemplateSection/index.tsx
